Add files via upload
All files contained herein are strictly prohibited from being used for any other purpose. Failure to comply with this directive will result in liability for criminal acts under the Republic Act No. 10175, otherwise known as the Cybercrime Prevention Act of 2012, and the Data Privacy Act of 2012 (Republic Act No. 10173).
 
This notice serves as a formal warning and reminder of the legal consequences of unauthorized access or use of these files.